--no-rm

+++ build.sh
@@ -5,9 +5,13 @@ if [ "$HB_INC_INSTALL" == "" ]; then
     exit -1
 fi
 
-rm -r -f .h
+if [ "$1" != "--no-rm" ] ; then
+   rm -r -f .h
+fi
 
 unset F18_DEBUG
+unset F18_DEBUG_BROWSE_SIF
+unset F18_DEBUG_FIN_AZUR
 
 ./build_lib.sh
